Rename Bruteforce Weak Key.py to Bruteforce_Weak_Key.py
Write a python code in cocalc that breaks the following using brute force attack:

Assume you got a ciphertext (in hexa) that  is  b'1525053514291239',  as a result of XOR encryption.
Additionally, assume that the unknown plaintext is an 8-letter word and among the most common words in English
You know that a very weak key is used that consists of a repeated character followed by @ sign. For example, if the word is an 8 letter word, then the key will be: ?@?@?@?@
Your task is to brute force this ciphertext based on the above information.
Output any successful decrypted plaintext, if you have found it among the most common words in English.